wip refactor aws4 signer utils
